backend/requirements.txt
altgraph==0.17.4
beautifulsoup4==4.12.3
blinker==1.8.2
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
Flask==3.0.3
Flask-Cors==5.0.0
Flask-MySQL==1.5.2
greenlet==3.1.1
idna==3.8
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.1
mysql-connector-python==9.0.0
packaging==24.1
pefile==2024.8.26
pyinstaller==6.10.0
pyinstaller-hooks-contrib==2024.8
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
pywin32-ctypes==0.2.3
requests==2.32.3
setuptools==74.0.0
six==1.16.0
soupsieve==2.6
SQLAlchemy==2.0.35
typing_extensions==4.12.2
urllib3==2.2.2
Werkzeug==3.0.4
